CHROMIUM: iwl7000: cfg80211: Add support for NAN Data Path (NDP) operations
Add support for request/respond/terminate an NDP with a peer
NAN device.

A request to start an NDP with a peer NAN device must be
issued on a NAN Data Interface (NDI) and must include the
cookie identifying the peer's publish function and other
optional parameters such as QoS and security configuration.
When the driver successfully processes the request it returns
a new NDP identifier that together with the NDI address
uniquely identifies the NDP. The actual NDP establishment
is performed asynchronously and must be completed with a
notification indicating the result of the session establishment.
The notification should be sent over the NAN Device interface.

A request to establish an NDP received from a peer is notified
over the NAN device interface with the relevant parameters
associated with the request, e.g., QoS and security information,
assuming that there is some user space process that would manage
the request. In case that the user space process decided to reject
the request, an NDP response with reject and reason code should be
issued. Otherwise, the user space process is expected to add a
NAN Data Interface (if needed) and send an NDP response requesting
the driver to complete the NDP establishment (over the NAN Data
Interface). As in the case of NDP request, the actual NDP
establishment is performed asynchronously and must be completed with
a notification indicating the result of the session establishment.
The notification should be sent over the NAN Device interface.

After an NDP is established it can be terminated by both the
initiator and the responder. A request to terminate an NDP on the
local device must be done over the NAN Device interface and must
include the NDI of the initiator and the NDP identifier.

NDP termination is notified over the NAN Device interface,
regardless if it was locally generated or not.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 06b81b2916094ce5788e58f8f07a2e164fa327f0