CHROMIUM: Bluetooth: mgmt: add additional HCI cmd to remove_advertising
If a conditional command is the last one in a request and is skipped,
bad thing occurs.

A work around is to append an additional HCI command in the tail of
the request command queue. This whatever HCI command triggers the
cleaning up work in the original hci event functions.

BUG=chromium:760792
TEST=Run magic tether flow and verify that bluetoothd hears back from
kernel for the completion of remove_advertising request.

Change-Id: If22999fe1dc28d03ba73eeebd5f3c46398c477c3
Signed-off-by: Joseph Hwang <josephsih@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/695947
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Commit-Queue: Miao-chen Chou <mcchou@chromium.org>
Tested-by: Miao-chen Chou <mcchou@chromium.org>