CHROMIUM: iwl7000: xvt: add common driver command
Add common driver command. This command is meant to serve as a wrapper
for commands with minimal implementation in the USC. It allows
the USC to serve merely as a pipe for tool<->driver/FW tests. It is
added as part of 'pyfi' tests requirements.

Signed-off-by: Nitzan Kugelman <nitzanx.kugelman@intel.com>
iwl7000-tree: 8caae1fa47fc75100470c837e5876dd5cb293776