FROMGIT: mwifiex: minor cleanups w/ sta_list_spinlock in cfg80211.c
The sta_list_spinlock looks to be used to control locking of the
list. Specifically when someone has the lock they may be allowed
to modify or delete elements of the list.

That implies that we shouldn't access the fields of the elements
returned by mwifiex_get_sta_entry() after we've released the
spinlock. Let's make some small changes so this is true.

It's unlikely that this matters since it looks to be just error
handling, but it's nice to be clean.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Ganapathi Bhat <gbhat@marvell.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from git://git.kernel.org/pub/scm/linux/kernel/git/kvalo/wireless-drivers-next.git master
 commit f0f7c2275fb925ea287e2da0585d17bde3d1fd2e)

BUG=b:35581110
TEST=kevin Wifi

Change-Id: I5ebfdc52ac9d390e894e3a1f09ae8c691dda12b9
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/394210