BACKPORT: KVM: LAPIC: Keep timer running when switching between one-shot and periodic mode
If we take TSC-deadline mode timer out of the picture, the Intel SDM
does not say that the timer is disable when the timer mode is change,
either from one-shot to periodic or vice versa.

After this patch, the timer is no longer disarmed on change of mode, so
the counter (TMCCT) keeps counting down.

So what does a write to LVTT changes ? On baremetal, the change of mode
is probably taken into account only when the counter reach 0. When this
happen, LVTT is use to figure out if the counter should restard counting
down from TMICT (so periodic mode) or stop counting (if one-shot mode).

This patch is based on observation of the behavior of the APIC timer on
baremetal as well as check that they does not go against the description
written in the Intel SDM.

Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: Radim Krm <rkrcmar@redhat.com>
Signed-off-by: Wanpeng Li <wanpeng.li@hotmail.com>
[Fixed rate limiting of periodic timer.]
Signed-off-by: Radim Krm <rkrcmar@redhat.com>

(cherry picked from commit dedf9c5e216902c6d34b5a0d0c40f4acbb3706d8)

Conflicts:
   arch/x86/kvm/lapic.c

[dtor: kvm_lapic_set_reg() is apic_set_reg() in 4.4; brought
limit_periodic_timer_frequency() and tests for type of timer from other
commits]

BUG=b:117265154
TEST=Build and boot, run crostini

Change-Id: I69e46b926de0da9a55902e3a51c8437a35998849
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1281925
Reviewed-by: Dylan Reid <dgreid@chromium.org>