BACKPORT: KVM: LAPIC: Apply change to TDCR right away to the timer
The description in the Intel SDM of how the divide configuration
register is used: "The APIC timer frequency will be the processor's bus
clock or core crystal clock frequency divided by the value specified in
the divide configuration register."

Observation of baremetal shown that when the TDCR is change, the TMCCT
does not change or make a big jump in value, but the rate at which it
count down change.

The patch update the emulation to APIC timer to so that a change to the
divide configuration would be reflected in the value of the counter and
when the next interrupt is triggered.

Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: Radim Krčmář <rkrcmar@redhat.com>
Signed-off-by: Wanpeng Li <wanpeng.li@hotmail.com>
[Fixed some whitespace and added a check for negative delta and running
 timer. - Radim]
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>

Conflicts:
   arch/x86/kvm/lapic.c

[delco: replaced remaining = 0; with remaining = ktime_set(0, 0);
since earlier commit to broadly make this type of change isn't
present]

BUG=b:120040011
TEST=Built

(cherry picked from commit c301b909e4b94e6395251787e28b335c51309fff)
Signed-off-by: Matt Delco <delco@chromium.org>
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I34410c0ff47f5f138236623bde52c4c5a84e22e2
Reviewed-on: https://chromium-review.googlesource.com/1766532
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>