CHROMIUM: mali: run dep_clear from jd_done to avoid deadlock
When jd_done_nolock/post_external_resources cleans up katom's reservation
callback and waits for worker to finish ctx lock is held, which the worker
needs to complete resulting in a deadlock.

We fix it here by using an atomic variable to synchronize with the worker and
running worker's payload from cleanup inside the locked section and converting
the worker to use mutex_trylock. So if the worker has not run its payload by the
time the katom is being cleaned up, the payload s going to run from cleanup and
the worker will do nothing.

BUG=b:35586182,chromium:770980
TEST=boot kevin with kernel build with USE=lockdebug

Change-Id: I7f67a6521fc59117e58bb61d9d49e329781120a0
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/706104
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>