CHROMIUM: netfilter: Downgrade netfilter warning that we're ignoring
Probably nobody is going to fix the warning talked about in
chromium:774279 and (as far as we know) it's not causing any serious
problems.  Let's downgrade it to just a normal warning printout
instead of a splat of a stackdump.  This way we don't get in the habit
of ignoring stackdumps that show up at bootup.

From the best I can tell, making a better fix would require us to pick
back a whole lot of patches and change APIs for all the netfilter
components.  Someone could do that if they were interested, but it
would take a bunch of careful work.

BUG=chromium:774279
TEST=No more splat at boot

Change-Id: I7fa6357ff2c2866ce748bf3d08dbed6e5b3f1342
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/716675
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>