CHROMIUM: arm64: dts: rockchip: scarlet: Reset Bluetooth at boot
Make sure the Bluetooth part of the QCA6174A is in a defined state
after boot by resetting it. Unfortunately there isn't a standard API
for this, so we use a device tree hack similar to CL:593409 for WiFi.
At boot the BT_EN pin is first set to low through pinctrl and later
to high through a fake regulator configured as supply of the regulator
for the WLAN part of the QCA6174A.

BUG=b:66922012
TEST=dmesg | grep rampatch
  => Bluetooth: hci0: using rampatch file: qca/rampatch_usb_00000302.bin
  bluetoothctl
    => shows controller

Change-Id: If3ce95053baf96685bc6a5a542616e8747849af8
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/724727
Reviewed-by: Brian Norris <briannorris@chromium.org>