Revert "CHROMIUM: atkbd: Enable wake events"
This reverts commit e6463eba1e46 ("CHROMIUM: atkbd: Enable wake events")

As discussed with dtor, this commit was put in 3.4 to fix some issues on
a particular platform (lumpy). That platform is not using 4.4, so we do
not need this anymore. This was never upstreamed. Also, now the upstream
has a better commit that implements this in the right way:

commit f13b2065de81 ("Input: i8042 - allow KBD and AUX ports to wake up
from suspend-to-idle")

The reason for this revert is that the next patch enables a log on wakeup
to identify the wake source, which will be helpful to debug S0ix system
(On S0ix systems, there is often no good way to tell the wakeup reason)
in the suspend failures, or spurious wake evetns.

Signed-off-by: Rajat Jain <rajatja@google.com>

BUG=None
TEST=Build, boot and use keyboard to see no issues (Eve).

Change-Id: I462557e41a87480dd3a6805edbf29a776dac2aa1
Reviewed-on: https://chromium-review.googlesource.com/724234
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>