CHROMIUM: Move some of the netfilter-related modules to in-kernel
Some of the iptables-restore scripts performed by Android expect some
netfilter modules to be already loaded. For instance, the script:

  *mangle
  -A tetherctrl_mangle_FORWARD -p tcp --tcp-flags SYN SYN \
    -j TCPMSS --clamp-mss-to-pmtu
  COMMIT

First checks if the "mangle" table is already loaded[1]. If it isn't,
it tries to load the "ip6_tables" ko[2] and checks for the table again
(since the Android config has all the netfilter modules built-in)[3]
without explicitly trying to load the relevant ko ("ip6table_mangle").
This causes the whole operation to fail without attempting to actually
run the command (which would have resulted in the kernel loading the
kernel by itself).

To avoid the above issue, and to also avoid sprinkling `modprobe`
commands all around the system (since Android runs in a container
without the CAP_SYS_MODULE capability), this change moves the
netfilter-related modules from being as a module to being built
in-kernel if they:

* Appear in the android-base.cfg list[4].
* Have the CONFIG_NETFILTER_XT_*, CONFIG_IP_NF*, CONFIG_IP6_NF_*
  prefixes.
* Are not explicitly loaded through other means (which exclude
  CONFIG_IP_NF_NAT, CONFIG_IP_NF_TARGET_MASQUERADE,
  CONFIG_IP6_NF_FILTER).

1:
https://android.googlesource.com/platform/external/iptables/+/master/\
iptables/ip6tables-restore.c#297
2:
https://android.googlesource.com/platform/external/iptables/+/master/\
libxtables/xtables.c#418
3:
https://android.googlesource.com/platform/external/iptables/+/master/\
iptables/ip6tables-restore.c#67
4:
https://android.googlesource.com/kernel/configs/+/master/\
android-4.4/android-base.cfg

BUG=b:68277526
TEST=The above iptables-restore script now succeeds
TEST=third_party/autotest/files/site_utils/bootperf-bin/bootperf
     seconds_power_on_to_login:
       before: 9.61 (avg, over 10 runs)
       after: 9.587 (avg, over 10 runs)
Signed-off-by: Luis Hector Chavez <lhchavez@chromium.org>
Change-Id: Ib58e9f4e080dd48fbacc7608655fea31e9e333e0
Reviewed-on: https://chromium-review.googlesource.com/747003
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>