CHROMIUM: bluetooth: properly handle changing values in conditional commands
This patch adopts the recipe from dmitrygr@ to fix the skip condition
for conditional commands as

    (cur_enabled == desired_enabled && !cur_changing) ||
    (cur_enabled != desired_enabled && cur_changing)

This skip condition works well for up to two HCI commands in progress.
This is good enough per bluetooth specification.

Debug logs are also added to monitor the number of commands in progress.

BUG=chromium:781749
TEST=Observe that the count of change_in_progress is not greater than 2
$ tail -f /var/log/messages | grep change_in_progress

Change-Id: I2ee2343aeda220da8d048472c33c33aadd50c487
Signed-off-by: Joseph Hwang <josephsih@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/756874
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Dmitry Grinberg <dmitrygr@google.com>