CHROMIUM: tpm: cr50_i2c: handle large burstcnt
Before this change, when reading the response header, we always read
burstcnt bytes and didn't check that the burstcnt is smaller than
the provided buffer. If the chip returned garbage in burstcnt,
we could read more than the provided buffer and corrupt the stack.

This CL adds an extra check and fails the read operation, if the
burstcnt is too large.

BUG=b:67923075
BUG=b:68143437
TEST=introduce artificial error, check that we don't read if
     burstcnt is garbage and bigger than the buffer size.
Signed-off-by: Andrey Pronin <apronin@chromium.org>

Change-Id: I03e6bc50d6927c1baf9edf4f7868c876907b6fe3
Reviewed-on: https://chromium-review.googlesource.com/765104
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>