CHROMIUM: device_jail: use a drop_privileges function for lockdown
The chromium-specific version of USBDEVFS_DROP_PRIVILEGES didn't
take any arguments and just set the privileges_dropped bit on the
usb_dev_state, so we were able to just do the ioctl from jail_device
directly. However, the upstream variant also takes a bitset that is
copied in from userspace, so this will no longer work. Instead we
can expose a function which performs the equivalent operation to
the old ioctl and call that from jail_device.

BUG=chromium:784936
TEST=using device_jail_server; security_DeviceJail_Lockdown passes

Change-Id: I8d8c901f0b3d429fc45962d45d951b881f99f0bc
Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/833172