FIXUP: CHROMIUM: iio: accel: Add cros_ec_accel_legacy driver
Because we read the LPC memory directly, we would read the primary EC
while setting up the downstream EC, incorrectly re-discovering the
accelerometers connected to the primary EC and associating them with the
secondary EC.

BUG=chromium:709642
TEST=On Samus with 4.4:
Before on the secondary EC (cros_pd), at
/sys/devices/.../PNP0C09:00/GOOG0004:00/cros-ec-ctl.2.auto/
We have chromeos, cros-ec-accel-legacy.0, cros-ec-accel-legacy.1
and cros-usb-pd-charger.
In iio/devices, we have cros-ec-accel and cros-ec-accel-legacy pointing
to the same accelerometer.

Now, only we have only chromeos and cros-usb-pd-charger.

Change-Id: I5f9b8c026ee0a9700d7e31a1c46357763c17b634
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/772807
Reviewed-by: Guenter Roeck <groeck@chromium.org>