CHROMIUM: drm/i915/cnl: Avoid DPLL 0 for HDMI to avoid Hard Hang.
XXX: This is a temporary hack to avoid Hard Hang while
we investigate the reasons behind the bug.

For some strange reason when disabling eDP (pipe A / DPLL 0)
and moving HDMI from pipe B with DPLL 1 to pipe A with DPLL0
the platform hard hang.

So for now, let's avoid DPLL 0 to avoid the Hard Hang.

v2: Nathan: Added the workaround for DP output as well

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>

BUG=chromium:783368
TEST=CNL boots to graphics

Change-Id: Ieb15d5511ef60182a83bc30a5b806bdf14898b88
Reviewed-on: https://chromium-review.googlesource.com/773781
Commit-Ready: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>