CHROMIUM: ASoC: nau8825: fix low volume of headphone.
nau8825_xtalk_baktab is used to backup and restore
registers in crosstalk. The values are 0 by default and
initialized in backup. Suppose a device has not plugged
a headphone after boot, which means backup has never been
called. After suspend/resume, restore will be called and
0 will be written to the registers. Both nau8825_xtalk_baktab
and resgisters will be 0. The headhpone volume will be
stuck low. Add a variable to indiciate if nau8825_xtalk_baktab
has been initialized.

BUG=chromium:769446
BUG=b:69087090
TEST=On caroline,
     (1) reboot (2) login (3) suspend/resume
     (4) plug a low impedance headphone (5) play youtube

Change-Id: I33a4c880ce434d1c3901923b8bb4e033bf4cb467
Signed-off-by: Wu-Cheng Li <wuchengli@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/778501
Reviewed-by: John Hsu <kchsu0@nuvoton.corp-partner.google.com>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>