FROMLIST: drm: add support for DisplayPort CEC-Tunneling-over-AUX
This adds support for the DisplayPort CEC-Tunneling-over-AUX
feature that is part of the DisplayPort 1.3 standard.

Unfortunately, not all DisplayPort/USB-C to HDMI adapters with a
chip that has this capability actually hook up the CEC pin, so
even though a CEC device is created, it may not actually work.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
(am from https://patchwork.freedesktop.org/patch/189282/)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Conflicts:
	include/drm/drm_dp_helper.h
	- rebase against v4.4 version

Change-Id: Icccfcaa036b49170cb09361ae4107a851346aed2
Reviewed-on: https://chromium-review.googlesource.com/1021371
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>