FROMLIST: drm/i915: add DisplayPort CEC-Tunneling-over-AUX support
Implement support for this DisplayPort feature.

The cec device is created whenever it detects an adapter that
has this feature. It is only removed when a new adapter is connected
that does not support this. If a new adapter is connected that has
different properties than the previous one, then the old cec device is
unregistered and a new one is registered to replace the old one.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
(am from https://patchwork.freedesktop.org/patch/189281/)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Conflicts:
	drivers/gpu/drm/i915/intel_dp.c
	- rebase against v4.4 version

Change-Id: Ia9580fe94b3dff665ece21c979ae3930fba3a9dd
Reviewed-on: https://chromium-review.googlesource.com/1021373
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>