CHROMIUMOS: config: disable uevent helper
We've enabled this historically but set the default path to "" and
nothing at runtime changed that which meant it wasn't actually used
anywhere.  Instead, we rely on standard udev processing netlink
messages and udev rules for hotplug behavior.

This knob can be used to subvert overall security at runtime by
changing /proc/sys/kernel/hotplug to point to arbitrary programs.

BUG=chromium:787016
TEST=precq passes

Change-Id: Id296f616edbee557bcc6f11fb343a17a019afc45
Signed-off-by: Mike Frysinger <vapier@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/779640
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>