CHROMIUM: esdfs: drop dentries on d_delete
There was a deadlock in the existing code to deal with
multiple views, as the inode lock is already held when
looking over the other superblocks. If this happens on
the same inode on two different views, they will deadlock.

Instead of cleaning up other views on rmdir/unlink, always
drop the dentries when they aren't in active use.

BUG=b:63876697
TEST=compilation, mount esdfs

Change-Id: If860b4d3bf8d4492c6f78c0564b6f4f2a7bb285d
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/887839
Commit-Ready: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>